Zmiana cen na Decimal

+++ FakTest/Sprzedaż.xaml.cs
@@ -78,9 +78,9 @@ namespace FakTest
                 produkt.typ = "brak";
                 produkt.netto = linia.cena.ToString() + "zł";
                 produkt.stawka = linia.VAT.ToString() + "%";
-                double temp = (linia.cena * linia.VAT/100);
+                decimal temp = (linia.cena * linia.VAT/100);
                 produkt.podatek = temp.ToString()  + " zł";
-                double temp2 = linia.cena + temp;
+                decimal temp2 = linia.cena + temp;
                 produkt.brutto = temp2.ToString() + " zł";
                 dataGrid.Items.Add(produkt);
             }
@@ -103,9 +103,9 @@ namespace FakTest
                 produkt.typ = "brak";
                 produkt.netto = linia.cena.ToString() + "zł";
                 produkt.stawka = linia.VAT.ToString() + "%";
-                double temp = (linia.cena * linia.VAT / 100);
+                decimal temp = (linia.cena * linia.VAT / 100);
                 produkt.podatek = temp.ToString() + " zł";
-                double temp2 = linia.cena + temp;
+                decimal temp2 = linia.cena + temp;
                 produkt.brutto = temp2.ToString() + " zł";
                 dataGrid.Items.Add(produkt);
             }
